refactor: extract canonical SSRF resolver
