+++ src/souwen/common_runtime/security/__init__.py
@@ -1,3 +1,5 @@
 """Security runtime boundary. Owner: Common Runtime. Allowed dependencies: standard library and common runtime."""
 
-__all__: list[str] = []
+from .fetch_target import ResolvedFetchTarget, resolve_fetch_target, validate_fetch_url
+
+__all__ = ["ResolvedFetchTarget", "resolve_fetch_target", "validate_fetch_url"]
